+++ classes/video_converter.h
@@ -1,3 +1,5 @@
+// fconvert v2.0.3 (c) 2023 - 2026 Eraldo Bako - MIT License
+// Maintaier: eraldobako@gmail.com
 #ifndef VIDEO_CONVERTER_H
 #define VIDEO_CONVERTER_H